add ability to add channels

+++ reed/Components/Settings/ChannelSettings/ChannelSettingsRow.swift
@@ -17,31 +17,35 @@ struct ChannelSettingsRow: View {
     var retrieveChannels: () -> Void
     
     var body: some View {
-        HStack {
-            Text(channel.title!).font(.headline)
-            Text(channel.link!)
-            Spacer()
-            Button(action: {
-                self.showingDeletionConfirmation = true
-            }, label: {
-                Text("􀈑")
-            }).alert(isPresented: $showingDeletionConfirmation) {
-                Alert(
-                    title: Text("Are you sure?"),
-                    message: Text("Are you sure you want to delete the channel '" + channel.title! + "'?"),
-                    primaryButton: .default(Text("Yes"), action: {() -> Void in
-                        viewContext.delete(channel)
-                        do {
-                            try context.save()
-                        } catch {
-                            print("Failed saving context after deleting!")
-                        }
-                        self.showingDeletionConfirmation = false
-                        retrieveChannels()
-                    }),
-                    secondaryButton: .default(Text("No"), action: {()->Void in
-                        self.showingDeletionConfirmation = false
-                    }))
+        if(channel.updateUri == nil) {
+            Text("Error")
+        } else {
+            HStack {
+                Text(channel.title!).font(.headline)
+                Text(channel.link!)
+                Spacer()
+                Button(action: {
+                    self.showingDeletionConfirmation = true
+                }, label: {
+                    Text("􀈑")
+                }).alert(isPresented: $showingDeletionConfirmation) {
+                    Alert(
+                        title: Text("Are you sure?"),
+                        message: Text("Are you sure you want to delete the channel '" + channel.title! + "'?"),
+                        primaryButton: .default(Text("Yes"), action: {() -> Void in
+                            viewContext.delete(channel)
+                            do {
+                                try viewContext.save()
+                            } catch {
+                                print("Failed saving context after deleting!")
+                            }
+                            self.showingDeletionConfirmation = false
+                            retrieveChannels()
+                        }),
+                        secondaryButton: .default(Text("No"), action: {()->Void in
+                            self.showingDeletionConfirmation = false
+                        }))
+                }
             }
         }
     }
